move CRUD actions into separate pageobject
