+++ tests/acceptance/config/behat.yml
@@ -192,6 +192,7 @@ default:
         - WebUIFilesContext:
         - WebUIGeneralContext:
         - WebUILoginContext:
+        - WebUISharingContext:
 
     webUIPersonalSettings:
       paths:
@@ -215,6 +216,7 @@ default:
         - WebUIFilesContext:
         - WebUIGeneralContext:
         - WebUILoginContext:
+        - WebUISharingContext:
 
     webUIRenameFolders:
       paths:
@@ -307,6 +309,7 @@ default:
         - WebUIFilesContext:
         - WebUIGeneralContext:
         - WebUILoginContext:
+        - WebUISharingContext:
 
   extensions:
       jarnaiz\JUnitFormatter\JUnitFormatterExtension:
