+++ tests/acceptance/features/lib/FilesPage.php
@@ -27,9 +27,8 @@ use Page\FilesPageElement\DetailsDialog;
 use Page\FilesPageElement\SharingDialog;
 use SensioLabs\Behat\PageObjectExtension\PageObject\Exception\ElementNotFoundException;
 use SensioLabs\Behat\PageObjectExtension\PageObject\Exception\UnexpectedPageException;
-use WebDriver\Exception\NoSuchElement;
-use WebDriver\Exception\StaleElementReference;
-use WebDriver\Key;
+use SensioLabs\Behat\PageObjectExtension\PageObject\Factory;
+use SensioLabs\Behat\PageObjectExtension\PageObject\Page;
 
 /**
  * Files page.
@@ -42,12 +41,13 @@ class FilesPage extends FilesPageBasic {
 	//see https://github.com/owncloud/core/issues/27870
 	protected $fileListXpath = ".//div[@id='app-content-files']//tbody[@id='fileList']";
 	protected $emptyContentXpath = ".//div[@id='app-content-files']//div[@id='emptycontent']";
-	protected $newFileFolderButtonXpath = './/*[@id="controls"]//a[@class="button new"]';
-	protected $newFolderButtonXpath = './/div[contains(@class, "newFileMenu")]//a[@data-templatename="New folder"]';
-	protected $newFolderNameInputLabel = 'New folder';
 	protected $newFolderTooltipXpath = './/div[contains(@class, "newFileMenu")]//div[@class="tooltip-inner"]';
-	protected $fileUploadInputId = "file_upload_start";
-	protected $uploadProgressbarLabelXpath = "//div[@id='uploadprogressbar']/em";
+	protected $deleteAllSelectedBtnXpath = ".//*[@id='app-content-files']//*[@class='delete-selected']";
+	/**
+	 *
+	 * @var FilesPageCRUD $filesPageCRUDFunctions
+	 */
+	protected $filesPageCRUDFunctions;
 	private $strForNormalFileName = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890';
 
 	/**
@@ -90,6 +90,25 @@ class FilesPage extends FilesPageBasic {
 	}
 
 	/**
+	 * @param Session $session
+	 * @param Factory $factory
+	 * @param array   $parameters
+	 */
+	public function __construct(
+		Session $session, Factory $factory, array $parameters = []
+	) {
+		parent::__construct($session, $factory, $parameters);
+		$this->filesPageCRUDFunctions = $this->getPage("FilesPageCRUD");
+		$this->filesPageCRUDFunctions->setXpath(
+			$this->emptyContentXpath,
+			$this->fileListXpath,
+			$this->fileNameMatchXpath,
+			$this->fileNamesXpath,
+			$this->deleteAllSelectedBtnXpath
+		);
+	}
+
+	/**
 	 * create a folder with the given name.
 	 * If name is not given a random one is chosen
 	 *
@@ -104,85 +123,9 @@ class FilesPage extends FilesPageBasic {
 		Session $session, $name = null,
 		$timeoutMsec = STANDARD_UI_WAIT_TIMEOUT_MILLISEC
 	) {
-		if ($name === null) {
-			$name = \substr(\str_shuffle($this->strForNormalFileName), 0, 8);
-		}
-
-		$newButtonElement = $this->find("xpath", $this->newFileFolderButtonXpath);
-
-		$this->assertElementNotNull(
-			$newButtonElement,
-			__METHOD__ .
-			" xpath $this->newFileFolderButtonXpath " .
-			"could not find new file-folder button"
+		return $this->filesPageCRUDFunctions->createFolder(
+			$session, $name, $timeoutMsec
 		);
-
-		$newButtonElement->click();
-
-		$newFolderButtonElement = $this->find("xpath", $this->newFolderButtonXpath);
-
-		$this->assertElementNotNull(
-			$newFolderButtonElement,
-			__METHOD__ .
-			" xpath $this->newFolderButtonXpath " .
-			"could not find new folder button"
-		);
-
-		try {
-			$newFolderButtonElement->click();
-		} catch (NoSuchElement $e) {
-			// Edge sometimes reports NoSuchElement even though we just found it.
-			// Log the event and continue, because maybe the button was clicked.
-			// TODO: Edge - if it keeps happening then find out why.
-			\error_log(
-				__METHOD__
-				. " NoSuchElement while doing newFolderButtonElement->click()"
-				. "\n-------------------------\n"
-				. $e->getMessage()
-				. "\n-------------------------\n"
-			);
-		}
-
-		try {
-			$this->fillField($this->newFolderNameInputLabel, $name . Key::ENTER);
-		} catch (NoSuchElement $e) {
-			// this seems to be a bug in MinkSelenium2Driver.
-			// Used to work fine in 1.3.1 but now throws this exception
-			// Actually all that we need does happen, so we just don't do anything
-		} catch (StaleElementReference $e) {
-			// At the end of processing setValue, MinkSelenium2Driver tries to blur
-			// away from the element. But we pressed enter which has already
-			// made the element go away. So we do not care about this exception.
-			// This issue started happening due to:
-			// https://github.com/minkphp/MinkSelenium2Driver/pull/286
-		}
-		$timeoutMsec = (int) $timeoutMsec;
-		$currentTime = \microtime(true);
-		$end = $currentTime + ($timeoutMsec / 1000);
-
-		while ($currentTime <= $end) {
-			$newFolderButton = $this->find("xpath", $this->newFolderButtonXpath);
-			if ($newFolderButton === null || !$newFolderButton->isVisible()) {
-				break;
-			}
-			\usleep(STANDARD_SLEEP_TIME_MICROSEC);
-			$currentTime = \microtime(true);
-		}
-		while ($currentTime <= $end) {
-			try {
-				$this->findFileRowByName($name, $session);
-				break;
-			} catch (ElementNotFoundException $e) {
-				//loop around
-			}
-			\usleep(STANDARD_SLEEP_TIME_MICROSEC);
-			$currentTime = \microtime(true);
-		}
-
-		if ($currentTime > $end) {
-			throw new \Exception("could not create folder");
-		}
-		return $name;
 	}
 
 	/**
@@ -209,16 +152,7 @@ class FilesPage extends FilesPageBasic {
 	 * @return void
 	 */
 	public function uploadFile(Session $session, $name) {
-		$uploadField = $this->findById($this->fileUploadInputId);
-		$this->assertElementNotNull(
-			$uploadField,
-			__METHOD__ .
-			" id $this->fileUploadInputId " .
-			"could not find file upload input field"
-		);
-		$uploadField->attachFile(\getenv("FILES_FOR_UPLOAD") . $name);
-		$this->waitForAjaxCallsToStartAndFinish($session, 20000);
-		$this->waitForUploadProgressbarToFinish();
+		$this->filesPageCRUDFunctions->uploadFile($session, $name);
 	}
 
 	/**
@@ -283,32 +217,39 @@ class FilesPage extends FilesPageBasic {
 		Session $session,
 		$maxRetries = STANDARD_RETRY_COUNT
 	) {
-		if (\is_array($toFileName)) {
-			$toFileName = \implode($toFileName);
-		}
+		$this->filesPageCRUDFunctions->renameFile(
+			$fromFileName, $toFileName, $session, $maxRetries
+		);
+	}
 
-		for ($counter = 0; $counter < $maxRetries; $counter++) {
-			try {
-				$fileRow = $this->findFileRowByName($fromFileName, $session);
-				$fileRow->rename($toFileName, $session);
-				break;
-			} catch (\Exception $e) {
-				$this->closeFileActionsMenu();
-				\error_log(
-					"Error while renaming file"
-					. "\n-------------------------\n"
-					. $e->getMessage()
-					. "\n-------------------------\n"
-				);
-			}
-		}
-		if ($counter > 0) {
-			$message = "INFORMATION: retried to rename file $counter times";
-			echo $message;
-			\error_log($message);
-		}
+	/**
+	 *
+	 * @param string|array $name
+	 * @param Session $session
+	 * @param bool $expectToDeleteFile
+	 * @param int $maxRetries
+	 *
+	 * @return void
+	 */
+	public function deleteFile(
+		$name,
+		Session $session,
+		$expectToDeleteFile = true,
+		$maxRetries = STANDARD_RETRY_COUNT
+	) {
+		$this->filesPageCRUDFunctions->deleteFile(
+			$name, $session, $expectToDeleteFile, $maxRetries
+		);
+	}
 
-		$this->waitTillFileRowsAreReady($session);
+	/**
+	 *
+	 * @param Session $session
+	 *
+	 * @return void
+	 */
+	public function deleteAllSelectedFiles(Session $session) {
+		$this->filesPageCRUDFunctions->deleteAllSelectedFiles($session);
 	}
 
 	/**
@@ -324,30 +265,9 @@ class FilesPage extends FilesPageBasic {
 	public function moveFileTo(
 		$name, $destination, Session $session, $maxRetries = STANDARD_RETRY_COUNT
 	) {
-		$toMoveFileRow = $this->findFileRowByName($name, $session);
-		$destinationFileRow = $this->findFileRowByName($destination, $session);
-
-		$this->initAjaxCounters($session);
-		$this->resetSumStartedAjaxRequests($session);
-		
-		for ($retryCounter = 0; $retryCounter < $maxRetries; $retryCounter++) {
-			$toMoveFileRow->findFileLink()->dragTo(
-				$destinationFileRow->findFileLink()
-			);
-			$this->waitForAjaxCallsToStartAndFinish($session);
-			$countXHRRequests = $this->getSumStartedAjaxRequests($session);
-			if ($countXHRRequests === 0) {
-				\error_log("Error while moving file");
-			} else {
-				break;
-			}
-		}
-		if ($retryCounter > 0) {
-			$message
-				= "INFORMATION: retried to move file $retryCounter times";
-			echo $message;
-			\error_log($message);
-		}
+		$this->filesPageCRUDFunctions->moveFileTo(
+			$name, $destination, $session, $maxRetries
+		);
 	}
 
 	/**
@@ -436,23 +356,6 @@ class FilesPage extends FilesPageBasic {
 	 * @return void
 	 */
 	public function waitForUploadProgressbarToFinish() {
-		$uploadProgressbar = $this->find(
-			"xpath", $this->uploadProgressbarLabelXpath
-		);
-		$this->assertElementNotNull(
-			$uploadProgressbar,
-			__METHOD__ .
-			" xpath $this->uploadProgressbarLabelXpath " .
-			"could not find upload progressbar"
-		);
-		$currentTime = \microtime(true);
-		$end = $currentTime + (STANDARD_UI_WAIT_TIMEOUT_MILLISEC / 1000);
-		while ($uploadProgressbar->isVisible()) {
-			if ($currentTime > $end) {
-				break;
-			}
-			\usleep(STANDARD_SLEEP_TIME_MICROSEC);
-			$currentTime = \microtime(true);
-		}
+		$this->filesPageCRUDFunctions->waitForUploadProgressbarToFinish();
 	}
 }
